Implemented Alternative Approach to visit each use case
Forest is now beeing invoked not only once and generates and smt
instance for all use cases.
The smt instances are collected and stored in a different database which
is permanent.
The smt instancens are then stored into one single SMT file.
At the moment a known problem ist that not all assertions together are
SAT. Due to that reason assert-soft is used NOW.